% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\docType{data}
\name{audio}
\alias{audio}
\title{Audio features from Stephen Breyer}
\format{
An object of class preppedAudio
}
\usage{
data(audio)
}
\description{
Features extracted from 100 randomly selected utterances by Stephen
Breyer in Supreme Court Oral Arguments.
}
\keyword{datasets}
