% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication-package.R
\docType{package}
\name{communication-package}
\alias{communication}
\alias{communication-package}
\title{An R package for Feature Extraction and Model Estimation for Audio of Human Speech}
\description{
Try the example below. Have a look at the references and learn more
from function documentation such as [hmm()].
}
\section{Options}{


TODO: think about package options, implement and describe them

- `communication.option1`: Should the output be coloured? (Default: `TRUE`).
- `communication.option2`: The maximum number of detailed test
   reports printed for the summary reporter (default: 10).

TODO: finalize it
}

\author{
\strong{Maintainer}: Christopher Lucas \email{christopher.lucas@wustl.edu}

Authors:
\itemize{
  \item Dean Knox \email{dcknox@upenn.edu}
}

Other contributors:
\itemize{
  \item Guilherme Duarte [contributor]
  \item Alex Shmuley [contributor]
  \item Vineet Bansal [contributor]
  \item Vadym Vashchenko [contributor]
}

}
\keyword{internal}
