% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{components}
\alias{components}
\alias{a_max}
\alias{zcr}
\alias{mcr}
\alias{max_min}
\alias{dc}
\alias{intensity}
\alias{loudness}
\alias{fastFourierTransform}
\alias{magPhase}
\alias{cacf}
\alias{melSpec}
\alias{mfcc}
\alias{formant}
\alias{vectorPreemphasis}
\alias{lpc}
\alias{delta}
\alias{delta2}
\alias{winFrame}
\alias{energy}
\alias{turnDetector}
\alias{addComponents}
\title{Components}
\usage{
a_max(config, input = "frames", output = "amax")

zcr(config, input = "frames", output = "zcr")

mcr(config, input = "frames", output = "mcr")

max_min(config, input = "frames", output = "maxmin")

dc(config, input = "frames", output = "dc")

intensity(config, input = "frames", output = "intensity")

loudness(config, input = "frames", output = "loudness")

fastFourierTransform(config, input = "frames", output = "fft")

magPhase(config, input = "fft", output = "magphase")

cacf(config, input = "magphase", output = "acf", usePower = 1, cepstrum = 0)

melSpec(
  config,
  input = "magphase",
  output = "melspec",
  usePower = 1,
  lofreq = 20,
  hifreq = 8000,
  nBands = 26
)

mfcc(
  config,
  input = "melspec",
  output = "mfcc",
  firstMfcc = 1,
  lastMfcc = 12,
  cepLifter = 22,
  htkcompatible = 1
)

formant(config, input = "lpc", output = "formants", nFormants = 3)

vectorPreemphasis(config, input = "frames", output = "pe", k = 0.97)

lpc(config, input = "pe", output = "lpc", method = "acf")

delta(config, input = "", output = NULL)

delta2(config, input = "", output = "dede")

winFrame(
  config,
  input = "frames",
  output = "winframes",
  winFunc = "ham",
  gain = 1,
  offset = 0
)

energy(config, input = "winframes", output = "energy", rms = 1, log_ = 0)

turnDetector(
  config,
  input = "energy",
  output = "turn",
  maxTurnLength = 12,
  maxTurnLengthGrace = 3,
  threshold = 0.0015
)

addComponents(config, func, input, output = NULL)
}
\arguments{
\item{config}{audio_config}

\item{input}{input}

\item{output}{output}

\item{usePower}{\code{integer}}

\item{cepstrum}{\code{binary}}

\item{lofreq}{\code{integer}}

\item{hifreq}{\code{integer}}

\item{nBands}{\code{integer}}

\item{firstMfcc}{\code{integer}}

\item{lastMfcc}{\code{integer}}

\item{cepLifter}{\code{double}}

\item{htkcompatible}{\code{integer}}

\item{nFormants}{\code{integer}}

\item{k}{\code{numeric} Coefficient for preemphasis}

\item{method}{\code{string} Methods 'act', 'burg'}

\item{winFunc}{\code{character}}

\item{gain}{\code{numeric}}

\item{offset}{\code{integer}}

\item{rms}{\code{integer}}

\item{log_}{\code{integer}}

\item{maxTurnLength}{\code{integer}}

\item{maxTurnLengthGrace}{\code{integer}}

\item{threshold}{\code{numeric}}

\item{func}{component function}
}
\description{
Describes components to be added to audio config 

In order to extract features from your data file, 
you can add components to your config
}
\section{Component types}{


Currently our package supports the following types of components.


\strong{Sample quantities} - Enable  quantities from sample value

\describe{
\item{`a_max`}{maximum absolute sample value}
\item{`zcr`}{zero-crossing rate}
\item{`mcr`}{mean-crossing rate}
\item{`maxmin`}{maximum and minimum sample value}
\item{`dc`}{dc-offset - arithmetic mean}
\item{`loudness`}{loudness}
\item{`intensity`}{intensity}
}
}

