% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.config.R
\name{createConfig}
\alias{createConfig}
\title{create audio_config}
\usage{
createConfig(nThreads = 1, frameSize = 0.025, frameStep = 0.0125)
}
\arguments{
\item{nThreads}{\code{integer}. defines the number of threads#'}

\item{frameSize}{\code{double}. defines the frame size}

\item{frameStep}{\code{double}. defines the frame step}
}
\value{
audio_config class
}
\description{
return config structure for extractFeatures
}
