% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.features.R
\name{extractFeatures}
\alias{extractFeatures}
\title{Extract features from a file and return audio class with these data}
\usage{
extractFeatures(filenames, config = loudness(createConfig()))
}
\arguments{
\item{filenames}{\code{character}. The list of paths and file names. For example, "folder/1.wav"}

\item{config}{\code{audio_config}. An object of class 'audio_config' with parameters for extraction.}
}
\value{
audio class
}
\description{
Extract features from a file and return audio class with these data
}
