% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.config.R
\name{generate_config_string}
\alias{generate_config_string}
\title{Transform audio.config to string}
\usage{
generate_config_string(audio_config)
}
\arguments{
\item{audio_config}{An audio_config object}
}
\value{
config string for openSmile
}
\description{
Transform audio.config to string
to be utilized in extract.features function
}
