% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.config.R
\name{getEdges}
\alias{getEdges}
\title{Method for obtaining the edges of an audio_config object}
\usage{
getEdges(x, ...)
}
\arguments{
\item{x}{audio object}

\item{...}{other parameters}
}
\value{
number of features
}
\description{
Method for obtaining the edges of an audio_config object
}
