% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.config.R
\name{getNumberFeatures}
\alias{getNumberFeatures}
\title{Method for obtaining the number of features to be extracted}
\usage{
getNumberFeatures(x, ...)
}
\arguments{
\item{x}{audio object}

\item{...}{other parameters}
}
\value{
number of features
}
\description{
Method for obtaining the number of features to be extracted
}
