% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llh.R
\name{llh}
\alias{llh}
\title{Title}
\usage{
llh(Xs, mod, control = list())
}
\arguments{
\item{Xs}{List of nsequences matrices; each matrix represents one observation
sequence and is of dimension nobs x nfeatures. For a single observation
sequence, a single matrix can be provided}

\item{mod}{Model object of class 'feelr.hmm', as output by \code{hmm}}

\item{control}{List of control parameters}
}
\value{
List with two components. \code{llhs} is a numeric vector of
  log-likelihoods of each observation sequence in \code{Xs}. \code{llh_total}
  is the log-likelihood of all observation sequences together, i.e.
  \code{sum(llhs)}. If \code{Xs} is the same data that generated \code{mod},
  the values calculated here will be slightly lower than those output in
  \code{mod$llhs}. This is because \code{hmm} estimates the starting state of
  each sequence, whereas here it is assumed that the starting state is drawn
  from the stationary distribution \code{mod$delta}.
}
\description{
Title
}
