% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-component.R
\name{pipeline.append}
\alias{pipeline.append}
\title{Add an individual component to an existing configuration object}
\usage{
pipeline.append(config, component, is_last = F)
}
\arguments{
\item{config}{A configuration object}

\item{component}{An individual component as a List object, with the first element
being the 'type' of the component (for example 'cTransformFFT').}

\item{is_last}{Whether this component is the last one to be added, in which case
its output channel will be set to the sink component, and the sink component's
input channel updated accordingly.}
}
\value{
A configuration object suitable for use with extractFeatures.
}
\description{
This function appends a (non-sink) component to an existing configuration object.
It is used internally by the pipeline function, and is unlikely to be used by
itself.
}
