% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-component.R
\name{pipeline}
\alias{pipeline}
\title{Construct a processing pipeline}
\usage{
pipeline(...)
}
\arguments{
\item{...}{Individual components specified as individual List objects.}
}
\value{
A configuration object suitable for use with extractFeatures.
}
\description{
This function creates a new Configuration composed of a pipeline of components.
Individual components of the pipeline are named automatically, but at a mininum
the type of each component (and any additional options needed for the component)
should be specified during construction.
}
\details{
Linear pipelines can be constructed easily by laying out the individual components
as individual List objects, where the first element of each List is the component
type (for example 'cWindower'). The input and output level names of each component
are determined automatically, but can be overridden, allowing for creation of 
more complex pipelines.
}
