% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.features.R
\name{play}
\alias{play}
\title{Play chunks of speech}
\usage{
play(x, ...)
}
\arguments{
\item{x}{An object of class 'speech'}

\item{...}{other parameters}
}
\value{
speech class
}
\description{
Play chunks of a speech object
}
