% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.utterances.R
\name{segment.utterances}
\alias{segment.utterances}
\title{Segment utterances by using voice activity detection}
\usage{
segment.utterances(
  folder = NULL,
  files = NULL,
  config = NULL,
  tsv.folder = NULL,
  use.full.names = F,
  use.exts = F
)
}
\arguments{
\item{folder}{Folder name of wav files to process}

\item{files}{Either a single wav file or a vector of wav files to process.
If specified, takes precedence over the folder argument}

\item{config}{A configuration object created through createConfig() or pipeline()}

\item{tsv.folder}{If specified, segment information of each input file is saved
to this folder in tsv format, with the same base name as the input file, but
with the .txt extension}

\item{use.full.names}{If TRUE, the complete path of each input file is used as
the name of the returned list components. If FALSE, only the base names are
used.}

\item{use.exts}{If TRUE, the extensions of the input files are used in the names
of the returned list components. If FALSE, the extensions are not used.}
}
\value{
A list of segment information, with names as the names of the input files
  encountered (but see use.full.names and use.exts arguments), and the values as
  a list of 2 vectors with name start and end, with the start and end times of
  each detected segment in seconds.
}
\description{
Segment utterances by using voice activity detection
}
