% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-params.R
\name{selectParams}
\alias{selectParams}
\title{Select hyper parameters for HMM model}
\usage{
selectParams(features, params = NULL, nfold = 5, cleanFeatures = T)
}
\arguments{
\item{features}{with labels}

\item{params}{parameters for CV - data.frame with columns nstates and lambda}

\item{nfold}{Number of folds}

\item{cleanFeatures}{clean features before running the model}
}
\value{
model object
}
\description{
Get features and return the hyperparameters with most likelihood
}
