% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.features.R
\name{labelUtterances}
\alias{labelUtterances}
\title{Label Utterances}
\usage{
labelUtterances(speeches, labels)
}
\arguments{
<<<<<<< HEAD
\item{speeches}{Vector of speeches.}
=======
\item{speeches}{Speeches}
>>>>>>> eadaa1871d3086e35138f278fc69e2c279d34b38

\item{labels}{\code{character}. Vector of labels for the speech object.}
}
\value{
speech object
}
\description{
<<<<<<< HEAD
Get a speech object and return it with labels
=======
Get speech objects and return them with labels
>>>>>>> eadaa1871d3086e35138f278fc69e2c279d34b38
}
