% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.features.R
\name{labelUtterances}
\alias{labelUtterances}
\title{Label Utterances}
\usage{
labelUtterances(speeches, labels)
}
\arguments{
\item{speeches}{Speeches}

\item{labels}{\code{character}. Vector of labels for the speech object.}
}
\value{
speech object
}
\description{
Get speech objects and return them with labels
}
