% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.features.R
\name{labelUtterances}
\alias{labelUtterances}
\title{Label Utterances}
\usage{
labelUtterances(features, labels)
}
\arguments{
\item{features}{Features object}

\item{labels}{\code{character}. Vector of labels for the speech object.}
}
\value{
speech object
}
\description{
Get a speech object and return it with labels
}
