% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.features.R
\name{play}
\alias{play}
\title{Play chunks of audio}
\usage{
play(x, ...)
}
\arguments{
\item{x}{An object of class 'audio'}

\item{...}{other parameters}
}
\value{
audio class
}
\description{
Play chunks of an audio object
}
